добавил блюпринт админ (#19)
