+++ webapp/templates/about.html
@@ -8,12 +8,12 @@
 <div class="container">
     <!-- Центральный блок -->
     <div class="px-4 py-5 my-5 text-center">
-        <h1 class="display-5 fw-bold">Centered hero</h1>
+        <h1 class="display-5 fw-bold">О нас</h1>
         <div class="col-lg-6 mx-auto">
-          <p class="lead mb-4">Quickly design and customize responsive mobile-first sites with Bootstrap, the world’s most popular front-end open source toolkit, featuring Sass variables and mixins, responsive grid system, extensive prebuilt components, and powerful JavaScript plugins.</p>
+          <p class="lead mb-4">Информация о проекте. Информация о проекте. Информация о проекте. Информация о проекте.</p>
           <div class="d-grid gap-2 d-sm-flex justify-content-sm-center">
-            <button type="button" class="btn btn-primary btn-lg px-4 gap-3">Primary button</button>
-            <button type="button" class="btn btn-outline-secondary btn-lg px-4">Secondary</button>
+            <button type="button" class="btn btn-primary btn-lg px-4 gap-3">Кнопка 1</button>
+            <button type="button" class="btn btn-outline-secondary btn-lg px-4">Кнопка 2</button>
           </div>
         </div>
     </div>
