+++ webapp/templates/base.html
@@ -14,27 +14,28 @@
       </a>
       <!-- menu -->
       <ul class="nav nav-pills">
-        <li class="nav-item"><a href="/" class="nav-link active" aria-current="page">Home</a></li>
+        <li class="nav-item"><a href="/" class="nav-link active" aria-current="page">Главная</a></li>
         <li class="nav-item"><a href="/users/login" class="nav-link">Авторизоваться</a></li>
         <li class="nav-item"><a href="/about" class="nav-link">О нас</a></li>
-        <li class="nav-item"><a href="#" class="nav-link">FAQs</a></li>
+        <li class="nav-item"><a href="#" class="nav-link">Пусто</a></li>
         <li class="nav-item"><a href="/users/logout" class="nav-link">logout</a></li>
       </ul>
      </header>
      </div>
   </head>
-
+<!--BODY-->
   <body>
     {% block body %}
     <div class="container"></div>
 
-<!--bootstrap-->
+    <!--bootstrap scripts-->
     <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js" integrity="sha384-oBqDVmMz9ATKxIep9tiCxS/Z9fNfEXiDAYTujMAeBAsjFuCZSmKbSSUnQlmh/jp3" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.min.js" integrity="sha384-cuYeSxntonz0PPNlHhBs68uyIAVpIIOZZ5JqeqvYYIcEL727kskC66kF92t6Xl2V" crossorigin="anonymous"></script>
     {% endblock %}
   </body>
+<!--End Body-->
 
-  <!--footer-->
+<!--footer-->
   <footer>
     <div class="container">
       <footer class="d-flex flex-wrap justify-content-between align-items-center py-3 my-4 border-top">
